Manual

Manual

Note

This package only works for optimization models that can be written in the conic-form.

Conic Duality

Conic duality is the starting point for MOI's duality conventions. When all functions are affine (or coordinate projections), and all constraint sets are closed convex cones, the model may be called a conic optimization problem. For conic-form minimization problems, the primal is:

\[\begin{align} & \min_{x \in \mathbb{R}^n} & a_0^T x + b_0 \\ & \;\;\text{s.t.} & A_i x + b_i & \in \mathcal{C}_i & i = 1 \ldots m \end{align}\]

and the dual is:

\[\begin{align} & \max_{y_1, \ldots, y_m} & -\sum_{i=1}^m b_i^T y_i + b_0 \\ & \;\;\text{s.t.} & a_0 - \sum_{i=1}^m A_i^T y_i & = 0 \\ & & y_i & \in \mathcal{C}_i^* & i = 1 \ldots m \end{align}\]

where each $\mathcal{C}_i$ is a closed convex cone and $\mathcal{C}_i^*$ is its dual cone.

For conic-form maximization problems, the primal is:

\[\begin{align} & \max_{x \in \mathbb{R}^n} & a_0^T x + b_0 \\ & \;\;\text{s.t.} & A_i x + b_i & \in \mathcal{C}_i & i = 1 \ldots m \end{align}\]

and the dual is:

\[\begin{align} & \min_{y_1, \ldots, y_m} & \sum_{i=1}^m b_i^T y_i + b_0 \\ & \;\;\text{s.t.} & a_0 + \sum_{i=1}^m A_i^T y_i & = 0 \\ & & y_i & \in \mathcal{C}_i^* & i = 1 \ldots m \end{align}\]

A linear inequality constraint $a^T x + b \ge c$ should be interpreted as $a^T x + b - c \in \mathbb{R}_+$, and similarly $a^T x + b \le c$ should be interpreted as $a^T x + b - c \in \mathbb{R}_-$. Variable-wise constraints should be interpreted as affine constraints with the appropriate identity mapping in place of $A_i$.

For the special case of minimization LPs, the MOI primal form can be stated as

\[\begin{align} & \min_{x \in \mathbb{R}^n} & a_0^T x &+ b_0 \\ & \;\;\text{s.t.} &A_1 x & \ge b_1\\ && A_2 x & \le b_2\\ && A_3 x & = b_3 \end{align}\]

By applying the stated transformations to conic form, taking the dual, and transforming back into linear inequality form, one obtains the following dual:

\[\begin{align} & \max_{y_1,y_2,y_3} & b_1^Ty_1 + b_2^Ty_2 + b_3^Ty_3 &+ b_0 \\ & \;\;\text{s.t.} &A_1^Ty_1 + A_2^Ty_2 + A_3^Ty_3 & = a_0\\ && y_1 &\ge 0\\ && y_2 &\le 0 \end{align}\]

For maximization LPs, the MOI primal form can be stated as:

\[\begin{align} & \max_{x \in \mathbb{R}^n} & a_0^T x &+ b_0 \\ & \;\;\text{s.t.} &A_1 x & \ge b_1\\ && A_2 x & \le b_2\\ && A_3 x & = b_3 \end{align}\]

and similarly, the dual is:

\[\begin{align} & \min_{y_1,y_2,y_3} & -b_1^Ty_1 - b_2^Ty_2 - b_3^Ty_3 &+ b_0 \\ & \;\;\text{s.t.} &A_1^Ty_1 + A_2^Ty_2 + A_3^Ty_3 & = -a_0\\ && y_1 &\ge 0\\ && y_2 &\le 0 \end{align}\]

Supported constraints

This is the list of supported Function-in-Set constraints of the package. If you try to dualize a constraint not listed here, it will return an usupported error.

MOI Function	MOI Set

SingleVariable	GreaterThan

SingleVariable	LessThan

SingleVariable	EqualTo

ScalarAffineFunction	GreaterThan

ScalarAffineFunction	LessThan

ScalarAffineFunction	EqualTo

VectorOfVariables	Nonnegatives

VectorOfVariables	Nonpositives

VectorOfVariables	Zeros

VectorOfVariables	SecondOrderCone

VectorOfVariables	RotatedSecondOrderCone

VectorOfVariables	PositiveSemidefiniteConeTriangle

VectorOfVariables	ExponentialCone

VectorOfVariables	DualExponentialCone

VectorOfVariables	PowerCone

VectorOfVariables	DualPowerCone

VectorAffineFunction	Nonnegatives

VectorAffineFunction	Nonpositives

VectorAffineFunction	Zeros

VectorAffineFunction	SecondOrderCone

VectorAffineFunction	RotatedSecondOrderCone

VectorAffineFunction	PositiveSemidefiniteConeTriangle

VectorAffineFunction	ExponentialCone

VectorAffineFunction	DualExponentialCone

VectorAffineFunction	PowerCone

VectorAffineFunction	DualPowerCone

Note that some of MOI constraints can be bridged, see Bridges, to constraints in this list.

Supported objective functions

MOI Function

SingleVariable

ScalarAffineFunction

Dualize a model

Dualization.dualize — Function.

dualize(args...; kwargs...)

The dualize function works in three different ways. The user can provide:

A MathOptInterface.ModelLike

The function will return a DualProblem struct that has the dualized model and PrimalDualMap{Float64} for users to identify the links between primal and dual model.

A MathOptInterface.ModelLike and a DualProblem{T}

A JuMP.Model

The function will return a JuMP model with the dual representation of the problem.

A JuMP.Model and an optimizer factory

The function will return a JuMP model with the dual representation of the problem with the OptimizerFactory attached. The OptimizerFactory is the solver and its key arguments that users provide in JuMP models, i.e. with_optimizer(GLPK.Optimizer).

On each of these methods, the user can provide the keyword argument dual_names. dual_names must be a DualNames struct. It allows users to set more intuitive names for the dual variables and dual constraints created.

source

DualOptimizer

You can solve a primal problem by using its dual formulation. ADD EXAMPLE.

Adding new sets

Dualization.jl can automatically dualize models with custom sets. To do this, the user needs to define the set and its dual set in MathOptInterface.jl and provide the functions dual_set and set_dot for the set the user wants to dualize.

ADD EXAMPLE