Reference

Reference

Dualization.dualize — Function.

dualize(model::MOI.ModelLike)

Dualize the model

source

Dualization.DualOptimizer — Type.

DualOptimizer(dual_optimizer::OT) where {OT <: MOI.ModelLike}

DualOptimizer constructor

source

Dualization.supported_constraints — Function.

supported_constraints(con_types::Vector{Tuple{DataType, DataType}})

Returns true if Function-in-Set is supported for Dualization and throws an error if it is not.

source

Dualization.supported_objective — Function.

supported_objective(primal_model::MOI.ModelLike)

Returns true if MOI.ObjectiveFunctionType() is supported for Dualization and throws an error if it is not.

source

Dualization.set_dual_model_sense — Function.

set_dualmodel_sense!(dual_model::MOI.ModelLike, model::MOI.ModelLike)

Set the dual model objective sense

source

Dualization.get_primal_objective — Function.

get_primal_obj_coeffs(model::MOI.ModelLike)

Get the coefficients from the primal objective function and return a PrimalObjectiveCoefficients{T}

source

Dualization.get_dual_objective — Function.

get_dual_obj_coeffs(dual_model::MOI.ModelLike, dict_constr_coeffs::Dict, 
                        dict_dualvar_primalcon::Dict, poc::POC{T}) where T

Get dual model objective function coefficients

source

Dualization.set_dual_objective — Function.

set_DOC(dual_model::MOI.ModelLike, doc::DualObjectiveCoefficients{T}) where T

Add the objective function to the dual model

source

Dualization.PrimalObjective — Type.

PrimalObjectiveCoefficients{T}

Primal objective coefficients defined as $a_0^Tx + b_0$ as in http://www.juliaopt.org/MathOptInterface.jl/stable/apimanual/#Advanced-1

source

Dualization.DualObjective — Type.

DualObjectiveCoefficients{T}

Dual objective coefficients defined as $b_i^Ty + b_0$ or $-b_i^Ty + b_0$ as in http://www.juliaopt.org/MathOptInterface.jl/stable/apimanual/#Advanced-1

source

Dualization.DualNames — Type.

DualNames

DualNames is a struct to pass the prefix of dual variables and dual constraints names. See more on naming the variables.

source

Dualization.dual_set — Function.

dual_set(s::MOI.AbstractSet) -> MOI.AbstractSet

Returns the dual set of s.

source