Reference

Dualization.supported_constraints — Function

supported_constraints(con_types::Vector{Tuple{DataType, DataType}})

Returns true if Function-in-Set is supported for Dualization and throws an error if it is not.

Dualization.supported_objective — Function

supported_objective(primal_model::MOI.ModelLike)

Returns true if MOI.ObjectiveFunctionType() is supported for Dualization and throws an error if it is not.

Dualization.set_dual_model_sense — Function

set_dual_model_sense!(dual_model::MOI.ModelLike, model::MOI.ModelLike)

Set the dual model objective sense.

Dualization.get_primal_objective — Function

get_primal_obj_coeffs(model::MOI.ModelLike)

Get the coefficients from the primal objective function and return a PrimalObjective{T}.

Dualization.get_dual_objective — Function

get_dual_objective(dual_model::MOI.ModelLike, dual_obj_affine_terms::Dict,
                   primal_objective::PrimalObjective{T})::DualObjective{T} where T

build the dual model objective function from the primal model.

Dualization.set_dual_objective — Function

set_dual_objective(dual_model::MOI.ModelLike, dual_objective::DualObjective{T})::Nothing where T

Add the objective function to the dual model.

Dualization.PrimalObjective — Type

PrimalObjective{T}

Primal objective is defined as a MOI.ScalarAffineFunction

Dualization.DualObjective — Type

DualObjective{T}

Dual objective is defined as a MOI.ScalarAffineFunction.

Dualization.DualNames — Type

DualNames

DualNames is a struct to pass the prefix of dual variables and dual constraints names. See more on naming the variables.

Missing docstring.

Missing docstring for Dualization.dual_set. Check Documenter's build log for details.