Reference

Dualization.supported_constraints — Function

supported_constraints(con_types::Vector{Tuple{DataType, DataType}})

Returns true if Function-in-Set is supported for Dualization and throws an error if it is not.

Dualization.supported_objective — Function

supported_objective(primal_model::MOI.ModelLike)

Returns true if MOI.ObjectiveFunctionType() is supported for Dualization and throws an error if it is not.

Dualization.DualNames — Type

DualNames

DualNames is a struct to pass the prefix of dual variables and dual constraints names. See more on naming the variables.